app/requirements.txt
asgiref==3.7.2
bottle==0.12.25
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
clr-loader==0.2.6
distlib==0.3.8
Django==5.0.3
event-emitter-js==0.1.2
filelock==3.15.4
greenlet==2.0.2
idna==3.6
json5==0.9.25
node-semver==0.9.0
nodeenv==1.8.0
peewee==3.17.6
pillow==10.2.0
platformdirs==4.2.2
playwright==1.35.0
playwright-stealth==1.0.6
proxy_tools==0.1.0
psutil==5.9.8
pycparser==2.22
pyee==9.0.4
pythonnet==3.0.3
pywebview==5.1
requests==2.31.0
segno==1.6.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.3
whatsapp-web==0.0.1
WPP_Whatsapp @ file:///E:/PythonProjects/DentalCareDev/WPP_Whatsapp-0.1.4.zip#sha256=2bd947cda988dbfb6033b7361632843c3437bd0f7b0d9b6831becc5d18c3cb74
